Renamed GOProperties to GOPropertiesDialog and moved to the dialogs subdirectoty
