+++ ide-projects/NetBeans12/nbproject/project.xml
@@ -6,7 +6,7 @@
             <name>GrandOrgue</name>
             <c-extensions>c</c-extensions>
             <cpp-extensions>cc,cpp,cxx</cpp-extensions>
-            <header-extensions>h,hpp,hxx,inc,linux</header-extensions>
+            <header-extensions>h,hpp,hxx,inl,linux</header-extensions>
             <sourceEncoding>UTF-8</sourceEncoding>
             <make-dep-projects/>
             <sourceRootList>
