+++ src/grandorgue/dialogs/GOPropertiesDialog.cpp
@@ -1,11 +1,11 @@
 /*
  * Copyright 2006 Milan Digital Audio LLC
- * Copyright 2009-2023 GrandOrgue contributors (see AUTHORS)
+ * Copyright 2009-2024 GrandOrgue contributors (see AUTHORS)
  * License GPL-2.0 or later
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
 
-#include "GOProperties.h"
+#include "GOPropertiesDialog.h"
 
 #include <wx/sizer.h>
 #include <wx/statline.h>
