+++ src/grandorgue/dialogs/GOPropertiesDialog.h
